test/requirements.txt
pytest==8.3.3
pytest-aiohttp==1.0.5
pytest-asyncio==0.24.0
